requirements.txt
asgiref==3.2.10
backports.entry-points-selectable==1.1.0
botocore==1.20.36
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
ClickUz==0.3
colorama==0.4.4
distlib==0.3.2
Django==3.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-adminlte-3==0.1.6
django-ckeditor==6.1.0
django-dynamic-preferences==1.11.0
django-js-asset==1.2.2
django-mptt==0.13.4
django-parler==2.2
django-payments==0.15.0
django-rosetta==0.9.7
django-storages==1.11.1
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
PaycomUz==3.0
persisting-theory==0.2.1
Pillow==8.3.2
platformdirs==2.3.0
polib==1.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.1
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.3
virtualenv==20.7.2
whitenoise==5.3.0
